% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from the ETL database}
\usage{
db_disconnect(con)
}
\arguments{
\item{con}{A \code{DBIConnection} object.}
}
\description{
Safely closes a DBI connection.
}
