% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epebdmc-package.R
\docType{package}
\name{epebdmc}
\alias{epebdmc-package}
\alias{epebdmc}
\title{epebdmc: Base de Dados de Mudanças Climáticas da EPE}
\description{
Fornece uma base de dados de indicadores socioambientais sobre riscos climáticos, mitigação e adaptação no setor de energia. Inclui ferramentas para consulta, análise e visualização dos indicadores de mudanças climáticas compilados pelo Sistema de Indicadores de Energia, Meio Ambiente e Sociedade (SIEMAS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abreucarvalho/epebdmc}
  \item Report bugs at \url{https://github.com/abreucarvalho/epebdmc/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Carvalho \email{rodrigo.carvalho@epe.gov.br}

}
