% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_to_csv}
\alias{load_to_csv}
\title{Load data to a CSV file}
\usage{
load_to_csv(data, path, ...)
}
\arguments{
\item{data}{A tibble/data.frame.}

\item{path}{Output file path.}

\item{...}{Additional arguments passed to \code{\link[readr:write_delim]{readr::write_csv()}}.}
}
\value{
Invisible file path.
}
\description{
Load data to a CSV file
}
