% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_enrich}
\alias{transform_enrich}
\title{Enrich data with derived columns}
\usage{
transform_enrich(data, ...)
}
\arguments{
\item{data}{A tibble/data.frame.}

\item{...}{Name-value pairs of new columns.}
}
\value{
An enriched tibble.
}
\description{
Add new computed columns using \code{\link[dplyr:mutate]{dplyr::mutate()}} syntax.
}
\examples{
\dontrun{
  data |> transform_enrich(
    emissionTon = emissionKt * 1000,
    decade      = (yearId \%/\% 10) * 10
  )
}
}
