% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_filter}
\alias{transform_filter}
\title{Filter data using expressions}
\usage{
transform_filter(data, ...)
}
\arguments{
\item{data}{A tibble/data.frame.}

\item{...}{Filter expressions (tidy eval).}
}
\value{
A filtered tibble.
}
\description{
Apply one or more filter conditions, just like \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\examples{
\dontrun{
  data |> transform_filter(ufId != 99, yearId >= 2010)
}
}
