% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_select}
\alias{transform_select}
\title{Select and rename columns}
\usage{
transform_select(data, ...)
}
\arguments{
\item{data}{A tibble/data.frame.}

\item{...}{Column selections (tidy-select).}
}
\value{
A tibble with selected columns.
}
\description{
Select and rename columns
}
