% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{uf_ibge_map}
\alias{uf_ibge_map}
\title{State name to IBGE code mapping}
\usage{
uf_ibge_map()
}
\value{
Named integer vector.
}
\description{
Returns a named integer vector mapping state names (as they appear
in the BTR1 spreadsheets) to official IBGE UF codes.
}
\keyword{internal}
